+++ doc/sphinx/src/news.rst
@@ -70,6 +70,7 @@ New features
 * Utilities
 
   * ``transform`` applies a transformation functor to a range.
+  * ``partition_point`` finds a value in a sorted range.
   * ``mpi::one_to_allv``, ``mpi::one_to_alli``, and ``mpi::all_to_allv`` additionally accept ranges and unary functors.
 
 * Logging